Update detail panel for consequentail second click on the samee quick jump.
When a user click on one of task IDs on the detail panel, it would
jump to and highlight the task; aka. quick jump.  With this change, by
clicking on the same quick jump twice, the detail panel would be
updated too.

For example, first click on SourceEventID would quick jump to the task
specified by SourceEventID.  If you click SourceEventID again, the
detail panel would be updated with the information of that task too.